.staging/realignment/website/pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = "egohygiene.io website"
authors = ["Universal <dev@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.12.0.88"
pillow = "^11.3.0"
readability-lxml = "^0.8.4.1"
beautifulsoup4 = "^4.13.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pre-commit-hooks = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.5"]
build-backend = "poetry.core.masonry.api"
